requirements.txt
flask_cors
python-dotenv
sqlalchemy
flask_sqlalchemy